更新redisClient以及通用工具类

+++ backend-manager/backend-manager-dao/src/main/java/cn/yutang/backend/dao/FoodOrdersCustomMapper.java
@@ -13,6 +13,6 @@ public interface FoodOrdersCustomMapper extends FoodOrdersMapper {
 
    List<FoodOrders> findOrders(@Param("page") Page page, @Param("orderCondition") OrderCondition orderCondition);
 
-   Long findCount(@Param("orderCondition") OrderCondition orderCondition);
+   Integer findCount(@Param("orderCondition") OrderCondition orderCondition);
 
 }
